pyproject.toml
[project]
name = "paella"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Dylan Dronnier", email = "dylan.dronnier@posteo.net" }
]
readme = "README.md"
requires-python = ">= 3.13"
dependencies = [
    "datasets[jax,vision]>=4.2.0",
    "flax>=0.12.0",
    "grain-nightly>=0.2.13.dev20251006",
    "hydra-zen>=0.16.0",
    "jax>=0.7.2",
    "optax>=0.2.6",
    "orbax>=0.1.9",
    "safetensors[jax,numpy]>=0.6.2",
    "tqdm>=4.67.1",
]

[project.scripts]
hyper = 'paella.experiments.hyperparams_opt:main'
train = 'paella.experiments.train_app:app'


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/paella"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "flagged",
    "mlruns",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    "I",
    "D",
    "E",    # pycodestyle errors.
    "F",    # Pyflakes rules.
    "PLC",  # Pylint convention.
    "PLC",  # Pylint convention warnings.
    "PLE",  # Pylint errors.
    "PLR",  # Pylint refactor recommendations.
    "PLW",  # Pylint warnings.
]

ignore = ["D203", "D213", "D211", "D407"]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = [
    "**/node_modules",
    "**/__pycache__",
    "**/*venv/**/*",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[dependency-groups]
dev = [
    "marimo[recommended]>=0.17.2",
    "plotly>=6.3.1",
    "pyrefly>=0.37.0",
    "ruff>=0.13.3",
]
